Static paths aren't going to do it.
This is because in `a <|> b`, `a`'s VDOM might change size
arbitrarily, and `b` has no way of knowing that.

One way to approach this is to assume that events are `mappend`ed
left-to-right and somehow keep the range of each `view` in every
trail, so that it can be effectively applied to the VDOM. Needs
more thought.